Unbundles circular dependency between ScriptParser and BabylonParser
