+++ packages/jest-editor-support/src/index.js
@@ -14,7 +14,8 @@ const Process = require('./Process');
 const ProjectWorkspace = require('./ProjectWorkspace');
 const Runner = require('./Runner');
 const Settings = require('./Settings');
-const {parse, Expect, ItBlock, Node} = require('./ScriptParser');
+const {Expect, ItBlock, Node} = require('./parsers/ParserNodes');
+const {parse} = require('./ScriptParser');
 const TestReconciler =  require('./TestReconciler');
 
 module.exports = {
